+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Increased incoming activity queue batch size to 20.
 - Change type of emoji reaction notification to `pleroma:emoji_reaction`.
 - Set limit on multipart form size.
+- Verify and drop portable activities sent to inbox.
 
 ### Removed